requirements.txt
#langchain==0.0.184
#PyPDF2==3.0.1
#python-dotenv==1.0.0
#streamlit==1.18.1
#faiss-cpu==1.7.4
#huggingface-hub==0.14.1
#InstructorEmbedding==1.0.1
#sentence-transformers==2.2.2

# ----

# streamlit
# python-dotenv
# PyPDF2
# langchain
# langchain-community
# langchain-groq
# chromadb
# sentence-transformers
# nomic

# --------------------------------------------------------------------

# Core Dependencies
streamlit>=1.30.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1

# PDF Processing (Multimodal)
unstructured[pdf]>=0.11.0
pdf2image>=1.16.0
pytesseract>=0.3.10
pdfplumber>=0.10.0
Pillow>=10.0.0
python-magic-bin>=0.4.14  # For Windows
PyMuPDF>=1.23.0

# LLM & Embeddings (100% FREE via Ollama)
langchain>=0.1.0
langchain-community>=0.0.38
ollama>=0.1.6

# Vector Database (FREE Tier - Qdrant Cloud)
qdrant-client>=1.7.0

# Utilities
dill>=0.3.7
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0
